requirements.txt
Flask
Pillow
requests
gunicorn
sentry-sdk[flask]